refactor(macos): remove frame invalidation plumbing
