+++ macos/Sources/Views/Agent/AgentContextBar.swift
@@ -13,7 +13,7 @@ public struct AgentContextBar: View {
     }
     public let state: AgentContextBarState
     @Environment(\.themeColors) private var theme
-    @Environment(\.guiFrameVersion) private var frameVersion
+
     public let encoder: InputEncoder?
 
     private let barHeight: CGFloat = 28
@@ -23,7 +23,6 @@ public struct AgentContextBar: View {
     @State private var elapsedDisplay: String = "0s"
 
     public var body: some View {
-        let _ = frameVersion
         if state.visible {
             HStack(spacing: 12) {
                 // Task description (left-aligned)
