+++ macos/Sources/Views/EditorChrome/StatusBarView.swift
@@ -46,7 +46,7 @@ public struct StatusBarView: View {
     public let state: StatusBarState
     public var feedbackState: FeedbackState?
     @Environment(\.themeColors) private var theme
-    @Environment(\.guiFrameVersion) private var frameVersion
+
     public let encoder: InputEncoder?
     public var isFileTreeVisible: Bool = false
     public var isGitStatusVisible: Bool = false
@@ -61,7 +61,6 @@ public struct StatusBarView: View {
     private let maxCenterStatusWidth: CGFloat = 320
 
     public var body: some View {
-        let _ = frameVersion
         GeometryReader { proxy in
             let layout = modelineLayout(totalWidth: proxy.size.width)
